[HW] Add Comment Attribute
Add a new CommentAttr attribute to the HW dialect.  This is a tighter
alternative than `sv::VerbatimOp` that allows a dialect to anchor string
comments to a specific operation.  This ensures that the comment is
always associated with the op to which it was attached regardless of how
the IR may be transformed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>